+++ resources/views/columns/checkbox.blade.php
@@ -2,4 +2,8 @@
 /** @var string $field */
 /** @var mixed $value */
 @endphp
-<input type="checkbox" class="form-control form-control-sm" name="{{ $field }}[]" value="{{ $value }}" role="grid-view-checkbox-item" />
+<div class="row">
+    <div class="col-12 text-center">
+        <input type="checkbox" class="form-control form-control-sm" name="{{ $field }}[]" value="{{ $value }}" role="grid-view-checkbox-item" />
+    </div>
+</div>