+++ changelog.md
@@ -1,5 +1,11 @@
 ### CHANGE LOG:
 
+**1.0.2 July 09, 2020:**
+- Ability to set filter's `name` by custom.
+- Add `sort` attribute and `getSort()` method to `BaseColumn`. Modify depended blocks.
+- Add `display` attribute to `CheckboxColumn` to set rule condition for displaying checkbox item.
+- Add `csrf_token` to `grid_view_rows_form`.
+
 **1.0.1 July 04, 2020:**
 - Readme fixes.
 - Change readme images to png.